IPC/BGA gen: Make output file handling explicit
Previously, which library a package ended up in depended on some
heuristics based on the footprint name, which could lead to footprints
from bga.yaml ending up in Package_CSP and vice versa. Now, we just have
a header in the yaml files that explicitly states what kind of package
this file contains, which determines where they end up.